Final Deliverables/NewsTracker/requirements.txt
flask
ibm_db
newsapi-python